requirements.txt
discord.py
python-dotenv
imgurpython
